+++ src/index.html
@@ -10,7 +10,7 @@
   <main class="container">
     <h1>Leaderboard</h1>
     <div class="score-container">
-      <h2>Recent scores</h2><button>Refresh</button>
+      <h2>Recent scores</h2><button class="refresh">Refresh</button>
       <ul class="score-list">
       </ul>
     </div>
