+++ src/modules/elements.js
@@ -1,0 +1,3 @@
+const scoreList = document.querySelector('.score-list');
+
+export { scoreList };
